third_party/WebKit/LayoutTests/http/tests/locks/workers.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Web Locks API: Workers</title>
<link rel=help href="https://github.com/inexorabletash/web-locks">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

let next_request_id = 0;
function postAndWait(worker, data) {
  return new Promise(resolve => {
    worker.postMessage(Object.assign(data, {rqid: next_request_id++}));
    const listener = event => {
      if (event.data.rqid !== data.rqid)
        return;
      worker.removeEventListener('message', listener);
      resolve(event.data);
    };
    worker.addEventListener('message', listener);
  });
}

async function postAndConfirm(frame, data) {
  const response = await postAndWait(frame, data);
  assert_equals(response.ack, data.op);
  assert_equals(response.id, data.lock_id);
  return response;
}

promise_test(async t => {
  const worker = new Worker('resources/worker.js');
  t.add_cleanup(() => { worker.terminate(); });

  const res = 'shared resource 1';

  await postAndConfirm(
    worker, {op: 'request', lock_id: 1, name: res, mode: 'shared'})

  await navigator.locks.acquire(res, {mode: 'shared'}, async lock => {
    await postAndConfirm(worker, {op: 'release', lock_id: 1});
  });

}, 'Window and Worker - shared mode');

promise_test(async t => {
  const worker = new Worker('resources/worker.js');
  t.add_cleanup(() => { worker.terminate(); });

  const res = 'exclusive resource 1';

  // worker acquires the lock.
  await postAndConfirm(worker, {op: 'request', lock_id: 1, name: res})

  // This request should be blocked.
  let lock_granted = false;
  const blocked = navigator.locks.acquire(
    res, lock => { lock_granted = true; });

  // Verify we can't get it.
  let available = undefined;
  await navigator.locks.acquire(
    res, {ifAvailable: true}, lock => { available = lock !== null; });
  assert_false(available);
  assert_false(lock_granted);

  // Ask the worker to release it.
  await postAndConfirm(worker, {op: 'release', lock_id: 1});

  // Now we've got it.
  const lock2 = await blocked;
  assert_true(lock_granted);

}, 'Window and Worker - exclusive mode');

promise_test(async t => {
  const worker1 = new Worker('resources/worker.js');
  const worker2 = new Worker('resources/worker.js');
  t.add_cleanup(() => { worker1.terminate(); worker2.terminate(); });

  const res = 'exclusive resource 2';

  // worker1 acquires the lock.
  await postAndConfirm(
    worker1, {op: 'request', lock_id: 1, name: res})

  // This request should be blocked.
  let lock_granted = false;
  const blocked = postAndConfirm(
    worker2, {op: 'request', lock_id: 1, name: res});
  blocked.then(f => { lock_granted = true; });

  // Verify worker2 can't get it.
  assert_true((await postAndConfirm(worker2, {
    op: 'request', name: res, lock_id: 2, ifAvailable: true
  })).failed, 'Lock request should have failed');
  assert_false(lock_granted);

  // Ask worker1 to release it.
  await postAndConfirm(worker1, {op: 'release', lock_id: 1});

  // Now worker2 can get it.
  const lock = await blocked;
  assert_true(lock_granted);

}, 'Worker and Worker - exclusive mode');

promise_test(async t => {
  const worker = new Worker('resources/worker.js');

  const res = 'exclusive resource 3';

  // Worker acquires the lock.
  await postAndConfirm(
    worker, {op: 'request', lock_id: 1, name: res})

  // This request should be blocked.
  let lock_granted = false;
  const blocked = navigator.locks.acquire(
    res, lock => { lock_granted = true; });

  // Verify we can't get it.
  let available = undefined;
  await navigator.locks.acquire(
    res, {ifAvailable: true}, lock => { available = lock !== null; });
  assert_false(available);
  assert_false(lock_granted);

  // Implicitly release it by terminating the worker.
  worker.terminate();

  // Now we've got it.
  const lock = await blocked;
  assert_true(lock_granted);

}, 'Terminated Worker - exclusive mode');

</script>
